+++ source/backend/cpu/CPUPool.cpp
@@ -404,7 +404,7 @@ ErrorCode CPUPool::onResize(const std::vector<Tensor *> &inputs, const std::vect
 
 ErrorCode CPUPool::onExecute(const std::vector<Tensor *> &inputs, const std::vector<Tensor *> &outputs) {
     MNN_CONCURRENCY_BEGIN(tId, mFunction.first) {
-        mFunction.second(tId);
+        mFunction.second((int)tId);
     }
     MNN_CONCURRENCY_END();
     return NO_ERROR;
